pyproject.toml
[project]
name = "clanker9000"
version = "0.1.0"
description = "SDK-first Discord bot with chat, shitposts, and voice pipeline."
authors = [
    { name = "alex furrier", email = "afurrier@gmail.com" }
]
requires-python = ">= 3.10"
readme = "README.md"
dependencies = [
    "aiohttp>=3.9.5",
    "discord.py>=2.4.0",
    "discord-ext-voice-recv>=0.4.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "ty>=0.0.2",        # Type checking
    "pytest>=8.1.1",    # Testing
    "pytest-cov>=5.0.0", # Coverage reporting
    "pytest-asyncio>=0.23.6",
    "dpytest>=0.7.0",
    "ruff>=0.3.0",      # Linting and formatting
    "pre-commit>=3.6.0", # Pre-commit hooks
    "tomli>=2.0.1",     # TOML reading
    "tomli-w>=1.0.0",   # TOML writing
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mkdocs-material>=9.6.14",
    "mkdocstrings[python]>=0.26.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src/clanker", "src/clanker_bot"]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings  
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "TID",    # flake8-tidy-imports
    "E402",   # module-import-not-at-top-of-file
    "N",      # pep8-naming
    "S",      # flake8-bandit
    "PTH",    # flake8-use-pathlib
    "RUF",    # ruff-specific rules
    "ICN",    # flake8-import-conventions
]

ignore = [
    "E501",   # line too long (handled by line-length)
    "TID252", # prefer absolute imports (keep relative for internal imports)
    "S101",   # assert used (acceptable in tests)
]

[tool.ruff.lint.flake8-tidy-imports]
# Prevent specific problematic imports
banned-api = {}

[tool.ruff.lint.isort]
known-first-party = ["clanker", "clanker_bot"]

[tool.ruff.lint.per-file-ignores]
"scripts/init_project.py" = ["I001", "PTH", "S602", "S603", "S607", "UP015"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "network: tests that require network access",
    "slow: slow tests",
]
asyncio_mode = "auto"
norecursedirs = ["scratch"]

[tool.ty.environment]
python-version = "3.9"

[tool.ty.src]
include = ["src"]

[tool.ty.terminal]
error-on-warning = true
